requirements.txt
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.2
cryptography==45.0.6
debugpy==1.8.13
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.2.0
firebase_admin==7.1.0
Flask==3.1.1
flask-cors==6.0.1
Flask-MySQL==1.6.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-genai==1.38.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.34.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
msgpack==1.1.1
mysql-connector-python==9.4.0
nest-asyncio==1.6.0
numpy==2.2.4
openai==1.108.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
pluggy==1.6.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-ffmpeg-video-streaming==0.1.16
pyzmq==26.4.0
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
tenacity==9.1.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.13.1
urllib3==2.5.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
